Move outfit landing to separate component
